Team "Factor" Models
Added a file 'single_game_reformatting.py' which converts the
training_set to a more practical version where each row is one game and
the columns are the various country viewerships. Added a file
'team_factor_models' which uses the one hot indicators of teams to
predict the total viewership, with the theory that each team has an
international following for their games.